Representation

BiochemicalVisualization.ball_and_stick

BiochemicalVisualization.stick

BiochemicalVisualization.van_der_waals

Available representations

BiochemicalVisualization.ball_and_stick — Function

ball_and_stick(::AbstractAtomContainer)

Creates and displays a ball-and-stick representation for the given atom container.

BiochemicalVisualization.stick — Function

stick(::AbstractAtomContainer)

Creates and displays a stick representation for the given atom container.

BiochemicalVisualization.van_der_waals — Function

van_der_waals(::AbstractAtomContainer)

Creates and displays a van-der-Waals representation for the given atom container. Sphere radii generally depend on the radius field of the corresponding atoms but are at least 1 Å.